:star: ht client trader changes default grid copy strategy to xls
